% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent.R
\name{as_matrix}
\alias{as_matrix}
\title{Convert Grid to Raster Matrix}
\usage{
as_matrix(grid)
}
\arguments{
\item{grid}{Grid object}
}
\value{
Numeric matrix
}
\description{
Extracts grid data as an R matrix.
}
\keyword{internal}
