% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_variable_ranges}
\alias{compute_variable_ranges}
\title{Compute Variable Ranges from Grids}
\usage{
compute_variable_ranges(grid_ptrs)
}
\arguments{
\item{grid_ptrs}{List of external pointers to Grid<float> objects.}
}
\value{
A data.frame with columns: min, max.
}
\description{
Scans all valid cells to determine min/max of each variable.
}
