% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coords_to_rowcol}
\alias{coords_to_rowcol}
\title{Convert geographic coordinates to row/col}
\usage{
coords_to_rowcol(dim_ptr, lon, lat)
}
\arguments{
\item{dim_ptr}{External pointer to GridDimension}

\item{lon}{Longitude}

\item{lat}{Latitude}
}
\value{
Integer vector with row and column indices
}
\description{
Convert geographic coordinates to row/col
}
