% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_grid_dimension}
\alias{create_grid_dimension}
\title{Create a GridDimension object}
\usage{
create_grid_dimension(nrows, ncols, xll, yll, cellsize)
}
\arguments{
\item{nrows}{Number of rows}

\item{ncols}{Number of columns}

\item{xll}{X coordinate of lower-left corner}

\item{yll}{Y coordinate of lower-left corner}

\item{cellsize}{Cell size (assumed square cells)}
}
\value{
External pointer to GridDimension object
}
\description{
Creates a grid dimension specifying the spatial extent and resolution
}
