% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_grid_float}
\alias{create_grid_float}
\title{Create a float Grid object}
\usage{
create_grid_float(dim_ptr, name, nodata_value = -9999)
}
\arguments{
\item{dim_ptr}{External pointer to GridDimension}

\item{name}{Grid layer name}

\item{nodata_value}{Value representing missing data}
}
\value{
External pointer to Grid<float> object
}
\description{
Creates a raster grid for environmental variables
}
