% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_layer}
\alias{create_layer}
\title{Create a Layer metadata object}
\usage{
create_layer(name, type_str)
}
\arguments{
\item{name}{Layer name.}

\item{type_str}{Type string: "Continuous", "Categorical", "Bias",
"Mask", "Probability", "Cumulative", "DebiasAvg", or "Unknown".}
}
\value{
External pointer to a Layer object.
}
\description{
Create a Layer metadata object
}
