% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_product_feature}
\alias{create_product_feature}
\title{Create a ProductFeature object}
\usage{
create_product_feature(values1, values2, name, min1, max1, min2, max2)
}
\arguments{
\item{values1}{Numeric vector for the first environmental variable}

\item{values2}{Numeric vector for the second environmental variable}

\item{name}{Feature name/identifier}

\item{min1}{Minimum of the first variable}

\item{max1}{Maximum of the first variable}

\item{min2}{Minimum of the second variable}

\item{max2}{Maximum of the second variable}
}
\value{
External pointer to ProductFeature object
}
\description{
Creates a product (interaction) feature between two environmental variables:
eval(i) = norm(values1[i]) * norm(values2[i]).
}
