% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_quadratic_feature}
\alias{create_quadratic_feature}
\title{Create a QuadraticFeature object}
\usage{
create_quadratic_feature(values, name, min_val, max_val)
}
\arguments{
\item{values}{Numeric vector of environmental variable values}

\item{name}{Feature name/identifier}

\item{min_val}{Minimum value for normalization}

\item{max_val}{Maximum value for normalization}
}
\value{
External pointer to QuadraticFeature object
}
\description{
Creates a quadratic feature: eval(i) = linear_val^2 where
linear_val = (values[i] - min) / (max - min).
}
