% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_sample}
\alias{create_sample}
\title{Create a Sample object}
\usage{
create_sample(point, row, col, lat, lon, name)
}
\arguments{
\item{point}{Point identifier}

\item{row}{Row index in grid}

\item{col}{Column index in grid}

\item{lat}{Latitude coordinate}

\item{lon}{Longitude coordinate}

\item{name}{Sample name/identifier}
}
\value{
External pointer to Sample object
}
\description{
Creates a species occurrence sample point
}
