% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_threshold_feature}
\alias{create_threshold_feature}
\title{Create a ThresholdFeature object}
\usage{
create_threshold_feature(values, name, threshold)
}
\arguments{
\item{values}{Numeric vector of environmental variable values}

\item{name}{Feature name/identifier}

\item{threshold}{The threshold value}
}
\value{
External pointer to ThresholdFeature object
}
\description{
Creates a binary step feature: eval(i) = 1.0 if values[i] > threshold, else 0.0.
}
