% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{csv_close}
\alias{csv_close}
\title{Close a CsvReader}
\usage{
csv_close(reader_ptr)
}
\arguments{
\item{reader_ptr}{External pointer to a CsvReader object.}
}
\value{
Called for side effects; returns invisibly.
}
\description{
Close a CsvReader
}
