% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{csv_headers}
\alias{csv_headers}
\title{Get CSV column headers}
\usage{
csv_headers(reader_ptr)
}
\arguments{
\item{reader_ptr}{External pointer to a CsvReader object.}
}
\value{
Character vector of header names.
}
\description{
Get CSV column headers
}
