% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{csv_open}
\alias{csv_open}
\title{Open a CSV file for reading}
\usage{
csv_open(filename, has_header = TRUE)
}
\arguments{
\item{filename}{Path to the CSV file.}

\item{has_header}{Logical: first line contains column names (default TRUE).}
}
\value{
External pointer to a CsvReader object.
}
\description{
Open a CSV file for reading
}
