% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{csv_read_double_column}
\alias{csv_read_double_column}
\title{Read an entire named column as doubles}
\usage{
csv_read_double_column(reader_ptr, field)
}
\arguments{
\item{reader_ptr}{External pointer to a CsvReader object.}

\item{field}{Column name.}
}
\value{
Numeric vector.
}
\description{
Reads from the current position to EOF.
}
