% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{csv_writer_close}
\alias{csv_writer_close}
\title{Close a CsvWriter}
\usage{
csv_writer_close(writer_ptr)
}
\arguments{
\item{writer_ptr}{External pointer to a CsvWriter object.}
}
\value{
Called for side effects; returns invisibly.
}
\description{
Close a CsvWriter
}
