% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{csv_writer_open}
\alias{csv_writer_open}
\title{Open a CSV file for writing}
\usage{
csv_writer_open(filename, append = FALSE, precision = 4L)
}
\arguments{
\item{filename}{Path to the output CSV file.}

\item{append}{Logical: append to existing file (default FALSE).}

\item{precision}{Number of decimal places for doubles (default 4).}
}
\value{
External pointer to a CsvWriter object.
}
\description{
Open a CSV file for writing
}
