% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{csv_writer_print_double}
\alias{csv_writer_print_double}
\title{Write a numeric value into the current row of a CSV}
\usage{
csv_writer_print_double(writer_ptr, column, value)
}
\arguments{
\item{writer_ptr}{External pointer to a CsvWriter object.}

\item{column}{Column name.}

\item{value}{Numeric value.}
}
\value{
Called for side effects; returns invisibly.
}
\description{
Write a numeric value into the current row of a CSV
}
