% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eval_auc}
\alias{eval_auc}
\title{Compute AUC (Area Under the ROC Curve)}
\usage{
eval_auc(presence, absence)
}
\arguments{
\item{presence}{Numeric vector of prediction scores at presence sites.}

\item{absence}{Numeric vector of prediction scores at absence sites.}
}
\value{
A named list with elements: auc, max_kappa, max_kappa_thresh.
}
\description{
Computes the Wilcoxon-Mann-Whitney AUC statistic from prediction scores
at presence and absence sites.
}
