% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eval_logloss}
\alias{eval_logloss}
\title{Compute Log-Loss}
\usage{
eval_logloss(presence, absence)
}
\arguments{
\item{presence}{Numeric vector of prediction scores at presence sites.}

\item{absence}{Numeric vector of prediction scores at absence sites.}
}
\value{
Average log-loss value.
}
\description{
Computes average cross-entropy log-loss from predictions at
presence and absence sites.
}
