% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eval_misclassification}
\alias{eval_misclassification}
\title{Compute Misclassification Rate}
\usage{
eval_misclassification(presence, absence)
}
\arguments{
\item{presence}{Numeric vector of prediction scores at presence sites.}

\item{absence}{Numeric vector of prediction scores at absence sites.}
}
\value{
Misclassification rate in [0, 1].
}
\description{
Fraction of misclassified samples at threshold 0.5.
}
