% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eval_square_error}
\alias{eval_square_error}
\title{Compute Mean Squared Error}
\usage{
eval_square_error(presence, absence)
}
\arguments{
\item{presence}{Numeric vector of prediction scores at presence sites.}

\item{absence}{Numeric vector of prediction scores at absence sites.}
}
\value{
Mean squared error.
}
\description{
Presence sites contribute (1-pred)^2, absence sites contribute pred^2.
}
