% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{feature_eval}
\alias{feature_eval}
\title{Evaluate a feature at a given index}
\usage{
feature_eval(feature_ptr, index)
}
\arguments{
\item{feature_ptr}{External pointer to a Feature object}

\item{index}{0-based index into the data vector}
}
\value{
Feature value at that index (double)
}
\description{
Evaluate a feature at a given index
}
