% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_grid_dimension_info}
\alias{get_grid_dimension_info}
\title{Get grid dimension properties}
\usage{
get_grid_dimension_info(dim_ptr)
}
\arguments{
\item{dim_ptr}{External pointer to GridDimension}
}
\value{
List with dimension properties
}
\description{
Get grid dimension properties
}
