% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_layer_info}
\alias{get_layer_info}
\title{Get layer metadata}
\usage{
get_layer_info(layer_ptr)
}
\arguments{
\item{layer_ptr}{External pointer to a Layer object.}
}
\value{
Named list with name and type (string).
}
\description{
Get layer metadata
}
