% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_sample_info}
\alias{get_sample_info}
\title{Get sample properties}
\usage{
get_sample_info(sample_ptr)
}
\arguments{
\item{sample_ptr}{External pointer to Sample}
}
\value{
List with sample properties
}
\description{
Get sample properties
}
