% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grid_float_from_matrix}
\alias{grid_float_from_matrix}
\title{Create a GridFloat from an R matrix}
\usage{
grid_float_from_matrix(mat, xll, yll, cellsize, nodata = -9999, name = "")
}
\arguments{
\item{mat}{Numeric matrix with grid data (NA becomes NODATA).}

\item{xll}{X coordinate of lower-left corner.}

\item{yll}{Y coordinate of lower-left corner.}

\item{cellsize}{Cell size.}

\item{nodata}{NODATA value (default -9999).}

\item{name}{Grid name (default "").}
}
\value{
External pointer to a GridFloat object.
}
\description{
Create a GridFloat from an R matrix
}
