% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grid_float_info}
\alias{grid_float_info}
\title{Get grid information for an ASC-loaded grid}
\usage{
grid_float_info(grid_ptr)
}
\arguments{
\item{grid_ptr}{External pointer to a GridFloat object.}
}
\value{
Named list with nrows, ncols, xll, yll, cellsize, nodata,
  name, count_data.
}
\description{
Returns metadata and basic statistics for a GridFloat.
}
