% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grid_float_to_matrix}
\alias{grid_float_to_matrix}
\title{Extract the data from a GridFloat as an R matrix}
\usage{
grid_float_to_matrix(grid_ptr)
}
\arguments{
\item{grid_ptr}{External pointer to a GridFloat object.}
}
\value{
Numeric matrix (nrows × ncols). NODATA cells are set to NA.
}
\description{
Extract the data from a GridFloat as an R matrix
}
