% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grid_from_matrix}
\alias{grid_from_matrix}
\title{Set grid from matrix}
\usage{
grid_from_matrix(grid_ptr, mat)
}
\arguments{
\item{grid_ptr}{External pointer to Grid}

\item{mat}{Numeric matrix of values}
}
\value{
Called for side effects; returns invisibly.
}
\description{
Set grid from matrix
}
