% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grid_get_value}
\alias{grid_get_value}
\title{Get grid value}
\usage{
grid_get_value(grid_ptr, row, col)
}
\arguments{
\item{grid_ptr}{External pointer to Grid}

\item{row}{Row index}

\item{col}{Column index}
}
\value{
Grid value
}
\description{
Get grid value
}
