% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent.R
\name{grid_info}
\alias{grid_info}
\title{Get Grid Information}
\usage{
grid_info(grid)
}
\arguments{
\item{grid}{Grid object}
}
\value{
List with grid properties
}
\description{
Retrieves properties of a grid object.
}
\keyword{internal}
