% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grid_read_asc}
\alias{grid_read_asc}
\title{Read an ESRI ASCII grid (.asc) file}
\usage{
grid_read_asc(filename)
}
\arguments{
\item{filename}{Path to the .asc file.}
}
\value{
External pointer to a GridFloat object.
}
\description{
Reads an ASC raster file and returns a float Grid external pointer.
}
