% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grid_read_file}
\alias{grid_read_file}
\title{Read a grid from file (auto-detect format)}
\usage{
grid_read_file(filename)
}
\arguments{
\item{filename}{Path to the grid file.}
}
\value{
External pointer to a GridFloat object.
}
\description{
Currently supports .asc format. Returns a float Grid.
}
