% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grid_set_value}
\alias{grid_set_value}
\title{Set grid value}
\usage{
grid_set_value(grid_ptr, row, col, value)
}
\arguments{
\item{grid_ptr}{External pointer to Grid}

\item{row}{Row index}

\item{col}{Column index}

\item{value}{Value to set}
}
\value{
Called for side effects; returns invisibly.
}
\description{
Set grid value
}
