% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grid_to_matrix}
\alias{grid_to_matrix}
\title{Get grid as matrix}
\usage{
grid_to_matrix(grid_ptr)
}
\arguments{
\item{grid_ptr}{External pointer to Grid}
}
\value{
Numeric matrix of grid values
}
\description{
Get grid as matrix
}
