% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{layer_name_from_path}
\alias{layer_name_from_path}
\title{Extract a layer name from a file path}
\usage{
layer_name_from_path(path)
}
\arguments{
\item{path}{File path.}
}
\value{
Character: layer name.
}
\description{
Strips directory and extension, e.g. "/data/bio1.asc" → "bio1".
}
