% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_io.R
\name{maxent_csv_close}
\alias{maxent_csv_close}
\title{Close a CSV Reader}
\usage{
maxent_csv_close(reader)
}
\arguments{
\item{reader}{External pointer to a CsvReader object.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Close a CSV Reader
}
