% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_io.R
\name{maxent_csv_headers}
\alias{maxent_csv_headers}
\title{Get CSV Column Headers}
\usage{
maxent_csv_headers(reader)
}
\arguments{
\item{reader}{External pointer to a CsvReader object.}
}
\value{
Character vector of column names.
}
\description{
Get CSV Column Headers
}
