% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maxent_featured_space_info}
\alias{maxent_featured_space_info}
\title{Get FeaturedSpace metadata}
\usage{
maxent_featured_space_info(fs_ptr)
}
\arguments{
\item{fs_ptr}{External pointer to a FeaturedSpace object.}
}
\value{
Named list with num_points, num_samples, num_features,
  density_normalizer, linear_predictor_normalizer, has_bias.
}
\description{
Get FeaturedSpace metadata
}
