% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maxent_get_entropy}
\alias{maxent_get_entropy}
\title{Get entropy of a FeaturedSpace distribution}
\usage{
maxent_get_entropy(fs_ptr)
}
\arguments{
\item{fs_ptr}{External pointer to a FeaturedSpace object.}
}
\value{
Shannon entropy (non-negative scalar).
}
\description{
Get entropy of a FeaturedSpace distribution
}
