% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maxent_get_loss}
\alias{maxent_get_loss}
\title{Get current loss of a FeaturedSpace}
\usage{
maxent_get_loss(fs_ptr)
}
\arguments{
\item{fs_ptr}{External pointer to a FeaturedSpace object.}
}
\value{
Scalar loss value (negative log-likelihood).
}
\description{
Get current loss of a FeaturedSpace
}
