% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maxent_get_weights}
\alias{maxent_get_weights}
\title{Get current distribution weights from a FeaturedSpace}
\usage{
maxent_get_weights(fs_ptr)
}
\arguments{
\item{fs_ptr}{External pointer to a FeaturedSpace object.}
}
\value{
Numeric vector of normalized weights (sums to 1).
}
\description{
Get current distribution weights from a FeaturedSpace
}
