% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent.R
\name{maxent_grid}
\alias{maxent_grid}
\title{Create a Grid Object}
\usage{
maxent_grid(dim, name = "", nodata_value = -9999)
}
\arguments{
\item{dim}{GridDimension object defining spatial extent}

\item{name}{Grid layer name}

\item{nodata_value}{Value representing missing data (default: -9999)}
}
\value{
Grid object (external pointer)
}
\description{
Creates a raster grid for storing environmental variable data.
}
\examples{
dim <- maxent_dimension(100, 100, -120, 35, 0.1)
grid <- maxent_grid(dim, "temperature")
}
