% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_io.R
\name{maxent_grid_info}
\alias{maxent_grid_info}
\title{Get Grid Information}
\usage{
maxent_grid_info(grid)
}
\arguments{
\item{grid}{External pointer to a GridFloat object.}
}
\value{
Named list with: \code{nrows}, \code{ncols}, \code{xll},
  \code{yll}, \code{cellsize}, \code{nodata}, \code{name},
  \code{count_data}.
}
\description{
Returns metadata about a GridFloat object read from a file.
}
