% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_io.R
\name{maxent_layer}
\alias{maxent_layer}
\title{Create a Layer Metadata Object}
\usage{
maxent_layer(name, type = "Continuous")
}
\arguments{
\item{name}{Character: layer name.}

\item{type}{Character: layer type. One of \code{"Continuous"},
\code{"Categorical"}, \code{"Bias"}, \code{"Mask"},
\code{"Probability"}, \code{"Cumulative"}, \code{"DebiasAvg"},
or \code{"Unknown"}.}
}
\value{
External pointer to a Layer C++ object.
}
\description{
Create a Layer Metadata Object
}
