% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_io.R
\name{maxent_layer_info}
\alias{maxent_layer_info}
\title{Get Layer Metadata}
\usage{
maxent_layer_info(layer)
}
\arguments{
\item{layer}{External pointer to a Layer object.}
}
\value{
Named list with \code{name} and \code{type}.
}
\description{
Get Layer Metadata
}
