% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featured_space.R
\name{maxent_model_entropy}
\alias{maxent_model_entropy}
\title{Get Model Entropy}
\usage{
maxent_model_entropy(featured_space)
}
\arguments{
\item{featured_space}{External pointer to a FeaturedSpace object.}
}
\value{
Scalar: Shannon entropy (non-negative).
}
\description{
Returns the Shannon entropy of the current MaxEnt distribution.
}
