% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featured_space.R
\name{maxent_model_loss}
\alias{maxent_model_loss}
\title{Get Model Loss}
\usage{
maxent_model_loss(featured_space)
}
\arguments{
\item{featured_space}{External pointer to a FeaturedSpace object.}
}
\value{
Scalar: current loss value.
}
\description{
Returns the current (negative log-likelihood) loss of the model.
}
