% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featured_space.R
\name{maxent_model_weights}
\alias{maxent_model_weights}
\title{Get Model Distribution Weights}
\usage{
maxent_model_weights(featured_space)
}
\arguments{
\item{featured_space}{External pointer to a FeaturedSpace object.}
}
\value{
Numeric vector of weights that sum to 1.
}
\description{
Returns the normalized probability weights of the current MaxEnt
distribution over the background points.
}
