% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{maxent_project_logistic_deprecated}
\alias{maxent_project_logistic_deprecated}
\title{Project Model onto Grids (Logistic Output, deprecated)}
\usage{
maxent_project_logistic_deprecated(model, env_grids, feature_names)
}
\arguments{
\item{model}{External pointer to a FeaturedSpace object.}

\item{env_grids}{List of external pointers to Grid<float> objects.}

\item{feature_names}{Character vector of environment variable names.}
}
\value{
External pointer to a Grid<float> with logistic scores.
}
\description{
logistic(x) = x / (1 + x). Produces values in [0, 1].
}
\keyword{internal}
