% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featured_space.R
\name{maxent_space_info}
\alias{maxent_space_info}
\title{Get FeaturedSpace Information}
\usage{
maxent_space_info(featured_space)
}
\arguments{
\item{featured_space}{External pointer to a FeaturedSpace object.}
}
\value{
Named list with: \code{num_points}, \code{num_samples},
  \code{num_features}, \code{density_normalizer},
  \code{linear_predictor_normalizer}.
}
\description{
Returns basic metadata about a FeaturedSpace object.
}
