% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent.R
\name{print.maxent_sample}
\alias{print.maxent_sample}
\title{Print Sample Information}
\usage{
\method{print}{maxent_sample}(x, ...)
}
\arguments{
\item{x}{Sample object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Sample Information
}
