% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{project_cloglog_deprecated}
\alias{project_cloglog_deprecated}
\title{Project Model onto Grids (cloglog output)}
\usage{
project_cloglog_deprecated(fs_ptr, grid_ptrs, feature_names)
}
\arguments{
\item{fs_ptr}{External pointer to a FeaturedSpace object.}

\item{grid_ptrs}{List of external pointers to Grid<float> objects.}

\item{feature_names}{Character vector of environment variable names.}
}
\value{
External pointer to a Grid<float> with cloglog scores in [0, 1].
}
\description{
cloglog(x) = 1 - exp(-x). Recommended output format for Maxent v3.4+.
}
