% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{project_logistic_deprecated}
\alias{project_logistic_deprecated}
\title{Project Model onto Grids (logistic output)}
\usage{
project_logistic_deprecated(fs_ptr, grid_ptrs, feature_names)
}
\arguments{
\item{fs_ptr}{External pointer to a FeaturedSpace object.}

\item{grid_ptrs}{List of external pointers to Grid<float> objects.}

\item{feature_names}{Character vector of environment variable names.}
}
\value{
External pointer to a Grid<float> with logistic scores in [0, 1].
}
\description{
logistic(x) = x / (1 + x).
}
