% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{project_raw}
\alias{project_raw}
\title{Project Model onto Grids (Java-compatible raw output)}
\usage{
project_raw(fs_ptr, grid_ptrs, feature_names)
}
\arguments{
\item{fs_ptr}{External pointer to a FeaturedSpace object.}

\item{grid_ptrs}{List of external pointers to Grid<float> objects.}

\item{feature_names}{Character vector of environment variable names.}
}
\value{
External pointer to a Grid<float> with Java raw scores.
}
\description{
Applies a trained FeaturedSpace model to produce Java Maxent "raw" scores:
  raw_java = exp(lp - lpNorm) / densityNorm
}
\details{
This matches the raw output of the Java Maxent software and the dismo
R package.
}
