% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{project_raw_deprecated}
\alias{project_raw_deprecated}
\title{Project Model onto Grids (raw output)}
\usage{
project_raw_deprecated(fs_ptr, grid_ptrs, feature_names)
}
\arguments{
\item{fs_ptr}{External pointer to a FeaturedSpace object.}

\item{grid_ptrs}{List of external pointers to Grid<float> objects.}

\item{feature_names}{Character vector of environment variable names,
matching the order of grid_ptrs.}
}
\value{
External pointer to a Grid<float> with raw prediction scores.
}
\description{
Applies a trained FeaturedSpace model to environmental grids to produce
raw Gibbs scores.
}
