% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_get_feature}
\alias{sample_get_feature}
\title{Get feature value from a sample}
\usage{
sample_get_feature(sample_ptr, feature_name, default_val = 0)
}
\arguments{
\item{sample_ptr}{External pointer to Sample}

\item{feature_name}{Name of the feature}

\item{default_val}{Default value if feature not found}
}
\value{
Feature value
}
\description{
Get feature value from a sample
}
