% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_set_feature}
\alias{sample_set_feature}
\title{Set feature value on a sample}
\usage{
sample_set_feature(sample_ptr, feature_name, value)
}
\arguments{
\item{sample_ptr}{External pointer to Sample}

\item{feature_name}{Name of the feature}

\item{value}{Feature value}
}
\value{
Called for side effects; returns invisibly.
}
\description{
Set feature value on a sample
}
