% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent.R
\name{set_grid_matrix}
\alias{set_grid_matrix}
\title{Set Grid from Matrix}
\usage{
set_grid_matrix(grid, mat)
}
\arguments{
\item{grid}{Grid object}

\item{mat}{Numeric matrix of values}
}
\value{
Invisibly returns the grid object.
}
\description{
Populates grid with values from an R matrix.
}
\keyword{internal}
