% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent.R
\name{maxent_dimension}
\alias{maxent_dimension}
\title{Create a Grid Dimension Object}
\usage{
maxent_dimension(nrows, ncols, xll, yll, cellsize)
}
\arguments{
\item{nrows}{Number of rows}

\item{ncols}{Number of columns}

\item{xll}{X coordinate of lower-left corner}

\item{yll}{Y coordinate of lower-left corner}

\item{cellsize}{Cell size (square cells)}
}
\value{
GridDimension object (external pointer)
}
\description{
Defines the spatial extent and resolution of a raster grid.
}
\examples{
dim <- maxent_dimension(
  nrows = 100, ncols = 100,
  xll = -120.0, yll = 35.0,
  cellsize = 0.1
)
}
