% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{maxent_feature_info}
\alias{maxent_feature_info}
\title{Get Feature Properties}
\usage{
maxent_feature_info(feature)
}
\arguments{
\item{feature}{External pointer to a Feature C++ object.}
}
\value{
Named list with elements: \code{name}, \code{type}, \code{lambda},
  \code{min}, \code{max}, \code{size}.
}
\description{
Returns a list of metadata for a feature object.
}
\examples{
vals <- c(0, 5, 10)
f <- maxent_linear_feature(vals, "temp")
maxent_feature_info(f)
}
