% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_io.R
\name{maxent_read_asc}
\alias{maxent_read_asc}
\title{Read an ESRI ASCII Grid File}
\usage{
maxent_read_asc(filename)
}
\arguments{
\item{filename}{Character: path to the \code{.asc} file.}
}
\value{
External pointer to a GridFloat C++ object.
}
\description{
Reads a \code{.asc} raster file into a GridFloat object.
}
\examples{
\donttest{
g <- maxent_read_asc("bio1.asc")
info <- maxent_grid_info(g)
print(info)
}
}
